change wait params to floats so you can sleep for partial seconds.  fix specs to reflect options params

+++ lib/helpers/wait.rb
@@ -2,8 +2,8 @@ module TestHelpers
   module Wait
     def poll_and_assert(options = {})
       return unless block_given?
-      timeout = Integer(options[:timeout]) || TestHelpers.configuration.wait_timeout
-      interval = Integer(options[:interval]) || TestHelpers.configuration.wait_interval
+      timeout = Float(options[:timeout] || TestHelpers.configuration.wait_timeout)
+      interval = Float(options[:interval] || TestHelpers.configuration.wait_interval)
       end_time = ::Time.now + timeout
       until ::Time.now > end_time
         begin
@@ -18,8 +18,8 @@ module TestHelpers
 
     def wait_until(options = {})
       return unless block_given?
-      timeout = Integer(options[:timeout]) || TestHelpers.configuration.wait_timeout
-      interval = Integer(options[:interval]) || TestHelpers.configuration.wait_interval
+      timeout = Float(options[:timeout] || TestHelpers.configuration.wait_timeout)
+      interval = Float(options[:interval] || TestHelpers.configuration.wait_interval)
       error = options[:error] || TimeoutError.new('Timed out waiting for block')
       end_time = ::Time.now + timeout
       until ::Time.now > end_time
